refactor: consolidate background manager
