+++ .changeset/background-agent-runtime.md
@@ -1,0 +1,6 @@
+---
+"@moonshot-ai/agent-core": patch
+"@moonshot-ai/kimi-code": patch
+---
+
+Consolidate background task management under the agent background runtime.
