+++ packages/agent-core/src/agent/background/process-task.ts
