+++ packages/agent-core/src/agent/background/task.ts
